Move native-contract macros into new crate soroban-native-sdk-macros
