+++ Cargo.toml
@@ -6,10 +6,12 @@ members = [
     "soroban-env-guest",
     "soroban-env-common",
     "soroban-env-macros",
+    "soroban-native-sdk-macros",
 ]
 
 [patch.crates-io]
 soroban-env-common = { path = "soroban-env-common" }
 soroban-env-macros = { path = "soroban-env-macros" }
+soroban-native-sdk-macros = { path = "soroban-native-sdk-macros" }
 stellar-xdr = { git = "https://github.com/stellar/rs-stellar-xdr", rev = "28a28dc0" }
 wasmi = { package = "soroban-wasmi", git = "https://github.com/stellar/wasmi", rev = "7b1f2355" }
