+++ soroban-env-host/Cargo.toml
@@ -10,7 +10,7 @@ readme = "../README.md"
 edition = "2021"
 
 [dependencies]
-soroban-env-macros = { version = "0.0.3" }
+soroban-native-sdk-macros = { version = "0.0.1" }
 soroban-env-common = { version = "0.0.3", features = [ "std" ] }
 wasmi = { package = "soroban-wasmi", version = "0.11.0", optional = true }
 static_assertions = "1.1.0"
