+++ soroban-env-host/src/native_contract/token/storage_types.rs
@@ -1,7 +1,7 @@
 use crate::native_contract::base_types::Map;
 use crate::native_contract::token::public_types::Identifier;
 use soroban_env_common::TryIntoVal;
-use soroban_env_macros::contracttype;
+use soroban_native_sdk_macros::contracttype;
 
 #[contracttype]
 pub struct AllowanceDataKey {
