+++ soroban-native-sdk-macros/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "soroban-native-sdk-macros"
+description = "Soroban native SDK macros."
+homepage = "https://github.com/stellar/rs-soroban-env"
+repository = "https://github.com/stellar/rs-soroban-env"
+authors = ["Stellar Development Foundation <info@stellar.org>"]
+license = "Apache-2.0"
+version = "0.0.1"
+edition = "2021"
+
+[lib]
+proc-macro = true
+
+[dependencies]
+syn = {version="1.0",features=["full"]}
+quote = "1.0"
+proc-macro2 = "1.0"
+itertools = "0.10.3"
+
+[package.metadata.docs.rs]
+all-features = true
