+++ soroban-native-sdk-macros/src/derive_fn.rs
